Move logging config to config file
